+++ packages/provider-utils/src/index.ts
@@ -31,6 +31,7 @@ export {
   asSchema,
   jsonSchema,
   lazySchema,
+  zodSchema,
   type FlexibleSchema,
   type InferSchema,
   type LazySchema,
@@ -53,7 +54,6 @@ export {
 export { VERSION } from './version';
 export { withUserAgentSuffix } from './with-user-agent-suffix';
 export * from './without-trailing-slash';
-export { zodSchema } from './zod-schema';
 
 // folder re-exports
 export * from './types';
