+++ packages/provider-utils/src/schema.test.ts
@@ -1,5 +1,5 @@
 import * as z from 'zod/v4';
-import { zodSchema } from './zod-schema';
+import { zodSchema } from './schema';
 import { safeParseJSON } from './parse-json';
 import { describe, expect, it } from 'vitest';
